[refactor] Session Ui 상태 타입 변경

+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/session/contract/uistate/MapUiState.kt
@@ -6,6 +6,6 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class MapUiState(
-    val paceColors: List<List<ULong>> = listOf(),
+    val paceColors: List<List<Int>> = listOf(),
     val path: List<List<LatLng>> = listOf(),
 ) : Parcelable
